Support REBOOT_READY broadcast
Support the new REBOOT_READY
broadcast by sending the broadcast with a positive
extra when the device becomes ready to reboot, and
by sending the broadcast with a negative extra if
it becomes active again.

Adds a new API that allows callers to cancel reboot
readiness checks, provided that they are the only
current client of the service.

Test: atest CtsRebootReadinessTestCases
Bug: 161353402
Change-Id: I57207cc84e5a8253fdcbaf61299e78fe7d6e80a6