Support reboot readiness DeviceConfig
Adds support for the new DeviceConfig namespace for
reboot readiness. These values are used to determine
the polling interval used when checking the device
state. These values are loaded at boot, and the
configurations are updated whenever a property
in the namespace is changed.

Added jarjar rules for handling DeviceConfig.

Test: atest CtsRebootReadinessTestCases
Bug: 161353402

Change-Id: I6e13ecb81d4788023da9247b8f45334f1f831469