Persist reboot readiness metrics
Stores information about reboot readiness events in the
module's system DE directory. This information will
be used in the future for logging metrics about automatic
reboots.

The data tracks the number of times each criterion blocked
the reboot, as well as the timestamp of when a pending
reboot was noted, and the device being ready to reboot.

Test: Mark pending reboot, wait for broadcast to be sent,
      inspect xml file
Bug: 161353402
Change-Id: I500c9c87dae78f51e81ed32ac70a64675155bf79