Add post reboot metrics collection
Adds metrics collection for unattended reboot
events by adding new SchedulingStatsLog class
for writing reboot readiness metrics to statsd.

The post reboot metrics are logged the first time
the USER_PRESENT broadcast is received after boot.
This is because the time taken for the first user
unlock is logged with this metric.

Adds host side test to test the end to end flow,
and ensure that the correct metrics are logged.

Test: atest RebootReadinessHostTest
Bug: 161353402

Change-Id: I4d9252448024b004f4c31599b511ffaecc5c4e40