Add tethering and alarm clock criteria
Adds new check that will block the reboot if there are any
active tethering clients.

Adds check that will block the reboot if there is an
alarm clock scheduled in the next 10 minutes (this
threshold is configurable). This check is used so that
the device will not reboot directly before an alarm, which
might mean the user would miss their alarm.

Also added DeviceConfig property for disabling subsystem
checks, while still allowing test callbacks to work. This
allows the CTS tests to just consider the test callbacks.

Test: atest CtsRebootReadinessTestCases
Test: Manual. Turn on tethering and disable other checks,
      note that device is not reboot ready when a second
      device is connected to the device's WiFi hotspot,
      and is reboot ready when the second device disconnects.
Test: Manual. Check that alarm set in the next 10 minutes
      stops the device from becoming reboot ready.
Bug: 161353402
Change-Id: I5408a35973deecaf32c8b5ed8e262fa2f855daf6