Schedule reboot readiness checks while in deep sleep
Instead of relying on a handler thread's scheduling, use
AlarmManager. This is because a handler thread is based on
the SystemClock uptime which does not increment during deep
sleep.

When scheduling reboot readiness checks, a new listener
is passed to AlarmManager. When this listener is
called, the reboot readiness checks will take place.

Test: Manual. Log timestamps of reboot readiness checks while
      the device is in a deep sleep state, and ensure they are
      consistent.
Bug: 185206044
Change-Id: Ia48586fdf52b2a4bb75dd2fad4340d1bd55bb2ad