Send broadcast immediately in multi-client case
If one client is already waiting for reboot readiness
broadcasts, send a broadcast to any new clients immediately
after they call markRebootPending. This is because the
new client would otherwise not know that the device is
already reboot-ready since broadcasts are usually only
sent when the state changes.

Also changes the broadcast logic so that explicit
broadcasts will only be sent to the user of the
original caller.

Test: atest CtsRebootReadinessTestCases
Test: atest RebootReadinessUnitTests
Bug: 185804886

Change-Id: I7b89961ea058b033fed17485f70bce8cb616ca3f