Change criteria for becoming not ready to reboot
Instead of polling the reboot readiness state while
idle, only poll the reboot readiness state if the
device interactivity changes while idle. This has
the benefit of consuming less resources while idle,
and also ensures that an app that starts in the
background while idle will not be able to influence
the reboot readiness signal.

This change negates the need for the idle polling
interval property. Several tests were changed to
account for this new behavior.

Test: atest CtsRebootReadinessTestCases
Test: atest RebootReadinessUnitTests
Test: atest RebootReadinessHostTest
Bug: 187736833
Change-Id: I7ad78b1e247d0a0b72721150a331c1bc11f6898b