Add reboot readiness shell command
Adds a new command line interface for interacting with
reboot readiness. These commands may be used to query
various criteria that make up the reboot signal, or to
listen for reboot readiness state changes for a set amount
of time.

Also adds additional logging to logcat and dumpsys in order
to better debug devices that are blocked.

Test: atest RebootReadinessUnitTests
Test: adb shell dumpsys reboot_readiness
Test: adb shell cmd reboot_readiness start-readiness-checks
Bug: 187407642
Change-Id: I8bab1065161b1c42ca0bb0dfe9747a97a0ee3e2b