Schedule state check before next alarm clock
If the device becomes ready to reboot, schedule a reboot
readiness check for when the next alarm clock is within
the reboot-blocking threshold, if the next alarm clock
exists. This will ensure that the alarm clock will
block the reboot if the device has remained ready
to reboot until then.

Test: atest RebootReadinessUnitTests
Bug: 190805342
Change-Id: I0e7c9325bee7ee17d206be210ec1f072f70289c5